Replace node server with mock-http objects
* Bring in mock-http so that we can use their mocks of http.ClientRequest and http.ServerResponse
as arguments to a RequestListener, which express implements
* Rework `proxy()` flow to accommodate this, while maintaining some semblance of API compatibility
* `createServer()` now returns the given request listener, which itself will have `_binaryTypes`.
This allows it to be supplied to `proxy()` instead of a full HTTP server. `serverListenCallback` will be invoked
immediately, given that there are no servers involved now and technically requests can be immediately processed
* `proxy()` now takes in the `binaryTypes` argument, so that one can just proxy the request listener, event, context
and optional binary types directly without prior set up
* Remove now-redundant functions relating to HTTP server lifecycle

Dispensing with the http server allows us to avoid the associated spin up time to open and attach to a socket, allowing us
to process requests much sooner.