example/lambda - init server once per context
The given example will create one http.Server _per lambda invocation_. This introduces significant delay, and for spiky traffic, would result in http.Server objects being thrown away when the excess lambda instances are removed.

This change attempts to improve things by storing an http.Server wrapped in a Promise in the execution context provided by AWS Lambda. The first lambda invocation will create the serverPromise and save it in the context. Subsequent invocations can then use this.

Given node assumes there is only a single thread of execution, it is obvious that there would be a race condition between two lambda invocations in creating and saving the serverPromise. This should only be a minor problem since the server itself is stateless and can be easily replaced without consequence.