set Content-Length header
- without content-length specified, it seems the body of DELETE requests get ignored
- per RFC 2616, section 4.3: The presence of a message-body in a request is signaled by the inclusion of a Content-Length or
Transfer-Encoding header field in the request's message-headers. A message-body MUST NOT be included in a request if
the specification of the request method (section 5.1.1) does not allow sending an entity-body in requests. A server SHOULD read and forward a message-body on any request; if the request method does not include defined semantics for an entity-body, then the message-body SHOULD be ignored when handling the request.
  - in other words, whether or not it is recommended to have a body in a DELETE request, if the content-length header (or the transfer-encoding header) is set, the message body should be forwarded
- concerns:
  - I'm not sure whether it is necessary to perform more type checking on `event.body`. for instance, if it is an object, it should be passed through a call to `JSON.stringify` before getting the byte length
  - the header should probably be set only if not already specified

This commit is one possible solution to awslabs/aws-serverless-express#106