Use nullish coalescing to determine path
We are currently struggling with a situation where `/test` proxies to a lambda. This causes `event.pathParameters.proxy` to be `undefined`. The request is then translated to `/test` rather than `/`.

To solve this we want to overwrite `event.pathParameters.proxy` with an empty string but the absence of nullish coalescing causes this to still use the `event.path` rather than `event.pathParameters.proxy`