fix: ensure alb response header values are strings (#699)
* fix: stringify response headers as ALB expects response headers to be type string

This was identified through using express within AWS Lambda which returns content-length header as a number which causes ALB to throw a 502 bad gateway exception. This change fixes this issue.

* fix: stringify response headers as ALB expects response headers to be type string

This was identified through using express within AWS Lambda which returns content-length header as a number which causes ALB to throw a 502 bad gateway exception. This change fixes this issue.