fix: fixing issue with async error handling within the SERVERLESS_EXPRESS:PROXY (#708)
* fix: stringify response headers as ALB expects response headers to be type string

This was identified through using express within AWS Lambda which returns content-length header as a number which causes ALB to throw a 502 bad gateway exception. This change fixes this issue.

* fix: stringify response headers as ALB expects response headers to be type string

This was identified through using express within AWS Lambda which returns content-length header as a number which causes ALB to throw a 502 bad gateway exception. This change fixes this issue.

* fix: fixing issue with async error handling within the SERVERLESS_EXPRESS:PROXY

This change handles non-async and async errors thrown within `forwardRequestToNodeServer` and adds an integration test (this was previously skipped).